Hoist page fetching so we can get the wpcom_performance_url object
